refactor near complete
