+++ .gitignore
@@ -1,7 +1,12 @@
 # do not push .sqlite files to GitHub
-**/*.sqlite
+app/*.sqlite
 __pycache__/
+app/__pycache__
+app/models/__pycache__
+app/routes/__pycache__
+app/db/__pycache__
 *.py[cod]
 auth
-app/__pycache__
 app/db.sqlite
+.github
+seeder.py
