+++ app/db/__init__.py
@@ -1,0 +1,3 @@
+from app.db.config import db
+import app.db.models as models
+
