+++ app/db/models/__init__.py
@@ -1,0 +1,9 @@
+from app.db import db
+from app.db.models.blog import BlogPost
+from app.db.models.contracts import Contracts
+from app.db.models.forum import ForumComment, ForumPost, Thread, ThreadRoles
+from app.db.models.customer import User, UserRoles, Role
+from app.db.models.customer import Company, CompanyMembers
+from app.db.models.customer import Order, ShippingAddress, Customer
+from app.db.models.product import Product
+from app.db.models.ticket import Ticket, TicketComment, TicketForum, Resolution
