+++ app/db/models/blog.py
@@ -1,5 +1,6 @@
-from .config import db
+from app.db import db
 from sqlalchemy import func
+
 
 class BlogPost(db.Model):
     __tablename__ = "blogpost"
