--- app/models/__init__.py
@@ -1,11 +1,0 @@
-# TODO Should we import all models into __init__.py for simplicity of importing models in the rest of the app?
-
-from .user import User, UserRoles, Role
-from .config import db
-from .blog import BlogPost
-from .company import Company, CompanyMembers
-from .contracts import  Contracts
-from .forum import ForumComment, ForumPost, Thread, ThreadRoles
-from .orders import Order, ShippingAddress, Customer
-from .product import Product
-from .ticket import Ticket, TicketComment, TicketForum, Resolution
