--- app/models/company.py
@@ -1,24 +1,0 @@
-from .config import db 
-
-
-class Company(db.Model):
-    __tablename__ = "company"
-    id = db.Column(db.Integer(), primary_key=True)
-    name = db.Column(db.String(50), unique=True)
-    phone = db.Column(db.String())
-    city = db.Column(db.String())
-    state = db.Column(db.String())
-    country = db.Column(db.String())
-    zipcode = db.Column(db.String())
-    email = db.Column(db.String(100), unique=True)
-    address1 = db.Column(db.String())
-    address2 = db.Column(db.String())
-    members = db.relationship('User', secondary='company_members')
-
-
-class CompanyMembers(db.Model):
-    __tablename__ = "company_members"
-    id = db.Column(db.Integer(), primary_key=True)
-    user_id = db.Column(db.Integer(), db.ForeignKey('user.id', ondelete='CASCADE'))
-    company_id = db.Column(db.Integer(), db.ForeignKey('company.id', ondelete='CASCADE'))
-    role_id = db.Column(db.Integer(), db.ForeignKey('roles.id', ondelete='CASCADE'))
