+++ app/routes/__init__.py
@@ -1,6 +1,6 @@
-from .forums import forums_blueprint
-from .auth import auth as auth_blueprint
-from .blogPosts import blogPosts as blogPosts_blueprint
-from .main import main as main_blueprint
-from .profiles import profiles as profile_blueprint
-from .profiles import company_info as company_blueprint
+from app.routes.main import main as main_blueprint
+from app.routes.auth import auth as auth_blueprint
+from app.routes.forums import forums_blueprint
+from app.routes.blogPosts import blogPosts as blogPosts_blueprint
+from app.routes.profiles import profiles as profile_blueprint
+from app.routes.profiles import company_info as company_blueprint
