+++ app/routes/auth.py
@@ -2,7 +2,9 @@ from flask import Blueprint, render_template, redirect, url_for, request, flash,
 from werkzeug.security import generate_password_hash, check_password_hash
 from flask_login import login_user, login_required, logout_user
 from flask_principal import identity_changed, Identity, AnonymousIdentity
-import app.models as model
+from app.db import db
+from app.db.models import User
+
 
 auth = Blueprint('auth', __name__)
 
@@ -15,15 +17,12 @@ def login_post():
     email = request.form.get('email')
     password = request.form.get('password')
     remember = True if request.form.get('remember') else False
-    user = model.User.query.filter_by(email=email).first()
-
+    user = User.query.filter_by(email=email).first()
     if not user or not check_password_hash(user.password, password):
         flash('Please check your login details and try again.')
         return redirect(url_for('auth.login')) 
-
     login_user(user, remember=remember)
     identity_changed.send(current_app._get_current_object(),identity=Identity(user.id))
-
     return redirect(url_for('profiles.profile'))
 
 @auth.route('/signup')
@@ -35,17 +34,14 @@ def signup_post():
     email = request.form.get('email')
     name = request.form.get('name')
     password = request.form.get('password')
-    user = model.User.query.filter_by(email=email).first() # if this returns a user, then the email already exists in database
-
+    user = User.query.filter_by(email=email).first() # if this returns a user, then the email already exists in database
     if user: # if a user is found, we want to redirect back to signup page so user can try again
         flash('Email address already exists')
 
         return redirect(url_for('auth.signup'))
-
-    new_user = model.User(email=email, name=name, password=generate_password_hash(password, method='sha256'))
-    model.db.session.add(new_user)
-    model.db.session.commit()
-
+    new_user = User(email=email, name=name, password=generate_password_hash(password, method='sha256'))
+    db.session.add(new_user)
+    db.session.commit()
     return redirect(url_for('auth.login'))
 
 @auth.route('/logout')
@@ -55,7 +51,5 @@ def logout():
     # Remove session keys set by Flask-Principal
     for key in ('identity.name', 'identity.auth_type'):
         session.pop(key, None)
-
     identity_changed.send(current_app._get_current_object(), identity=AnonymousIdentity())
-
     return redirect(url_for('main.home'))
