+++ app/routes/main.py
@@ -1,5 +1,6 @@
 from flask import Blueprint, render_template
 
+
 main = Blueprint('main', __name__)
 
 @main.route("/")
