+++ app/routes/profiles.py
@@ -1,8 +1,11 @@
 from flask import Blueprint, render_template, redirect, url_for, request
 from flask_login import login_required, current_user
-import app.models as models
+from app.db import db
+from app.db.models import User, Company, CompanyMembers
+
 
 profiles = Blueprint('profiles', __name__, url_prefix="/profile")
+company_info = Blueprint('company_info', __name__, url_prefix="/companies")
 
 @profiles.route('/')
 @login_required
@@ -15,12 +18,10 @@ def edit_profile_post():
     old_email = request.form.get('old_email')
     name = request.form.get('name')
     email = request.form.get('email')
-    user = models.User.query.filter_by(email=old_email).first()
-
+    user = User.query.filter_by(email=old_email).first()
     user.name = name
     user.email = email
-    models.db.session.commit()
-
+    db.session.commit()
     return redirect(url_for('profiles.profile'))
 
 @profiles.route('/edit_password')
@@ -30,13 +31,9 @@ def edit_password():
 @profiles.route('/edit_password', methods=['POST'])
 def edit_password_post():
     return redirect(url_for('profiles.profile'))
-    
-
-company_info = Blueprint('company_info', __name__, url_prefix="/companies")
 
 def get_companies():
-    return models.Company.query.all()
-
+    return Company.query.all()
 
 @company_info.route("/")
 def display_companies_home():
@@ -45,7 +42,7 @@ def display_companies_home():
 @company_info.route('/<int:company_id>')
 @login_required
 def display_company_info(company_id):
-    company = models.Company.query.get_or_404(company_id)
+    company = Company.query.get_or_404(company_id)
     return render_template('company_info/company_info.html', company=company)
 
 @company_info.route('/create', methods=['GET', 'POST'])
@@ -61,22 +58,19 @@ def create_company():
         country = request.form.get('country')
         phone = request.form.get('phone')
         email = request.form.get('email')
-
-        new_company = models.Company(name=name, address1=address1, address2=address2, city=city, state=state, zipcode=zipcode, country=country, phone=phone, email=email)
-        models.db.session.add(new_company)
-        models.db.session.commit()
-        new_members_company = models.CompanyMembers(id = new_company.id , user_id = current_user.id)
-        models.db.session.add(new_members_company)
-        models.db.session.commit()
-
+        new_company = Company(name=name, address1=address1, address2=address2, city=city, state=state, zipcode=zipcode, country=country, phone=phone, email=email)
+        db.session.add(new_company)
+        db.session.commit()
+        new_members_company = CompanyMembers(id = new_company.id , user_id = current_user.id)
+        db.session.add(new_members_company)
+        db.session.commit()
         return redirect(url_for("company_info.display_company_info", company_id=new_company.id))
-
     return render_template('company_info/company_info_create.html', companies=get_companies())
 
 @company_info.route('/edit_company_info/<int:company_id>', methods=['POST'])
 @login_required
 def edit_company_info_post(company_id):
-    company = models.Company.query.filter_by(id=company_id).first()
+    company = Company.query.filter_by(id=company_id).first()
     company.name = request.form.get('name')
     company.address1 = request.form.get('address1')
     company.address2 = request.form.get('address2')
@@ -86,6 +80,5 @@ def edit_company_info_post(company_id):
     company.country = request.form.get('country')
     company.phone = request.form.get('phone')
     company.email = request.form.get('email')
-    models.db.session.commit()
-
+    db.session.commit()
     return redirect(url_for('company_info.display_company_info', company_id=company.id))
